+++ src/records/Domain.php
@@ -8,6 +8,7 @@
 
 namespace flipbox\domains\records;
 
+use Craft;
 use flipbox\craft\sortable\associations\records\SortableAssociation;
 use flipbox\craft\sortable\associations\services\SortableAssociations;
 use flipbox\domains\db\DomainsQuery;
@@ -16,6 +17,8 @@ use flipbox\domains\fields\Domains;
 use flipbox\domains\validators\DomainValidator;
 use flipbox\ember\helpers\ModelHelper;
 use flipbox\ember\traits\SiteRules;
+use flipbox\ember\validators\LimitValidator;
+use yii\base\InvalidArgumentException;
 
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
@@ -95,6 +98,27 @@ class Domain extends SortableAssociation
                 ],
                 [
                     [
+                        'fieldId'
+                    ],
+                    LimitValidator::class,
+                    'query' => function (Domain $model) {
+                        return $model::find()
+                            ->field($model->fieldId)
+                            ->element($model->elementId)
+                            ->site($model->siteId)
+                            ->andWhere([
+                                '!=',
+                                static::TARGET_ATTRIBUTE,
+                                $model->{static::TARGET_ATTRIBUTE}
+                            ]);
+                    },
+                    'limit' => function (Domain $model) {
+                        return $this->getFieldLimit($model->fieldId);
+                    },
+                    'message' => "Limit exceeded."
+                ],
+                [
+                    [
                         'fieldId',
                         'status',
                     ],
@@ -105,5 +129,33 @@ class Domain extends SortableAssociation
                 ]
             ]
         );
+    }
+
+    /**
+     * @param int $fieldId
+     * @return Domains
+     */
+    protected function resolveField(int $fieldId): Domains
+    {
+        $field = Craft::$app->getFields()->getFieldById($fieldId);
+
+        if ($field === null || !$field instanceof Domains) {
+            throw new InvalidArgumentException(sprintf(
+                "Field must be an instance of '%s', '%s' given.",
+                Domains::class,
+                $field ? get_class($field) : 'FIELD NOT FOUND'
+            ));
+        }
+
+        return $field;
+    }
+
+    /**
+     * @param int $fieldId
+     * @return int
+     */
+    protected function getFieldLimit(int $fieldId): int
+    {
+        return (int) $this->resolveField($fieldId)->limit;
     }
 }